Reduce size of win streak in save block to u8 Added flag to handle IsFirstTypeWin Changed over all functions from Battle Tower to Restricted Sparring Added func to check Symbol eligibility

+++ include/global.h
@@ -367,7 +367,7 @@ struct BattleDomeTrainer
 struct RestrictedSparring
 {
     //u16 monId[FRONTIER_PARTY_SIZE-1];
-    u16 winStreak;
+    u8 winStreak;
 }__attribute__((packed));
 
 #define DOME_TOURNAMENT_TRAINERS_COUNT 16
